test(SDK-5354) - Reduce data classes
